Fix warning
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@27007 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/rcfpch/rpc_transport.c
@@ -254,7 +254,7 @@ rpc_transport_init()
 #else
     addr.sun_family = AF_UNIX;
     snprintf(addr.sun_path, sizeof(addr.sun_path), 
-             "/tmp/terpc_%ld", getpid());
+             "/tmp/terpc_%lu", (unsigned long)getpid());
     len = sizeof(addr) - UNIX_PATH_MAX + strlen(addr.sun_path);
 #endif
